Refactoring the cart-panel component and extracting item data.
