+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/shop/cart/cart-item/cart-item-sum.component.ts
@@ -1,5 +1,5 @@
 import { div } from '../../../framework/dom-creators';
-import { priceType } from './cart-item.types';
+import { priceType } from '../cart.types';
 
 export const cartItemSum = ({ value, currency = 'PLN' }: priceType): HTMLElement => {
   const $panelBlock = div('panel-block is-justify-content-end');
