+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/shop/cart/cart-panel/cart-panel.component.ts
@@ -1,17 +1,20 @@
-import { article, p } from '../framework/dom-creators.ts';
-import { cartItem } from './cart/cart-item/cart-item.component.ts';
-import { cartItemSum } from './cart/cart-item/cart-item-sum.component.ts';
+import { article, p } from '../../../framework/dom-creators';
+import { cartItem } from '../cart-item/cart-item.component';
+import { cartItemSum } from '../cart-item/cart-item-sum.component';
+import { cartPanelType } from '../cart.types';
 
-export function cartPanel({ heading = 'Cart Items', items }) {
+export const cartPanel = ({ heading = 'Cart Items', items }: cartPanelType): HTMLElement => {
   const $article = article('panel is-primary my-6 w-75 mx-auto');
   const $heading = p('panel-heading');
   $heading.textContent = heading;
   $article.append($heading);
   let value = 0;
+
   items.forEach((item) => {
     value += item.price.value;
     $article.appendChild(cartItem(item));
   });
+
   $article.append(cartItemSum({ value, currency: 'PLN' }));
   return $article;
-}
+};
