+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/shop/cart/cart.types.ts
@@ -9,3 +9,8 @@ export interface cartItemType {
   unit: 'kg';
   price: priceType;
 }
+
+export interface cartPanelType {
+  heading: string;
+  items: cartItemType[];
+}
