+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/shop/items.data.ts
@@ -1,0 +1,14 @@
+export const items = [
+    {
+        name: 'Tomatoes',
+        amount: 2,
+        unit: 'kg',
+        price: { value: 20, currency: 'PLN' }
+    },
+    {
+        name: 'Banana',
+        amount: 5,
+        unit: 'kg',
+        price: { value: 36, currency: 'PLN' }
+    }
+];
