+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/shop/root.component.js
@@ -1,21 +1,7 @@
 import { div } from '../framework/dom-creators.ts';
-import { cartPanel } from './cart-panel.component.js';
+import { cartPanel } from './cart/cart-panel/cart-panel.component.ts';
 import { hero } from './hero.component.js';
-
-const items = [
-  {
-    name: 'Tomatoes',
-    amount: 2,
-    unit: 'kg',
-    price: { value: 20, currency: 'PLN' }
-  },
-  {
-    name: 'Banana',
-    amount: 5,
-    unit: 'kg',
-    price: { value: 36, currency: 'PLN' }
-  }
-];
+import { items } from './items.data.ts';
 
 export function root() {
   const $hero = hero({ title: 'Shopping App', subTitle: 'buy and sell' });
